Widen ReliabilityBucket.n_observations/n_ambiguous to float
Task 5's nightly decay (0.95*old + 0.05*today) applied to n makes these
genuinely fractional after the first fold, but the Pydantic model still
typed them int -- model_validate() would raise on any real decayed row,
a latent crash for Task 6's get_risk. Confirmed by reviewer as a real
bug introduced by Task 5's own decay behavior.

Adds a regression test that folds two days with different observation
counts into the same bucket key (producing a genuinely fractional
n_observations, not a coincidental integer) and round-trips it through
ReliabilityBucket.model_validate() without raising. Verified the test
fails against the old int-typed field and passes after the widening.

+++ backend/app/models/risk.py
@@ -47,7 +47,11 @@ class ReliabilityBucket(BaseModel):
     hour_bucket: int
     stat_type: Literal["headway", "delay", "prediction_error"]
     histogram: dict
-    n_observations: int
-    n_ambiguous: int
+    # float, not int: Task 5's nightly decay (0.95*old + 0.05*today) applied
+    # to n the same way as the histogram makes these genuinely fractional
+    # after the first fold (e.g. 0.95), so a strict int field would raise
+    # ValidationError on real decayed rows.
+    n_observations: float
+    n_ambiguous: float
     window_start: date
     last_updated: datetime
